Moved local strategy to the authentication folder
